Move problem_file_line into utils
